refactor(cloister): extract workspace sync aggregation
The first item commit tripped the file-size ratchet before push. Extract the

new aggregation and lower the merge-agent baseline in the same item.
